Vec256Test was renamed as vec256_test_all_types
