+++ aten/src/ATen/test/CMakeLists.txt
@@ -84,7 +84,7 @@ list(APPEND ATen_MOBILE_TEST_SRCS
 
 list(APPEND ATen_VEC256_TEST_SRCS
   ${CMAKE_CURRENT_SOURCE_DIR}/vec256_test.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/Vec256Test.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/vec256_test_all_types.cpp
   )
 
 # ---[ Send the lists to the parent scope.
